pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-ai-mcp"
version = "0.1.0"
description = "Agentic AI application using MCP protocol and LangChain framework"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-community>=0.0.20",
    "anthropic>=0.18.0",
    "requests>=2.31.0",
    "duckduckgo-search>=4.0.0",
    "youtube-search-python>=1.6.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "websockets>=11.0.0",
    "beautifulsoup4>=4.12.0",
    "geopy>=2.3.0",
]

[tool.setuptools]
packages = ["src"]